Add support for selecting children
Children can now be selected using dot-style, bracket-style, or
index-style notations. All expressions must start with the `$` character
which indicates the root node of the JSON payload. Following the root
character can be any of the child notation styles chained together.

Dot-style notation is a `.` followed by a series of characters that make
up a _simple_ name. Simple names can be composed of any letter or digit,
or underscore. Some examples of this are `$.name` and `$.long_name`.
This notation selects the value of a JSON object by its name.

Bracket-style notation is a pair of `[]` with a series of characters
that make up a _complex_ name in between. Complex names can be composed
of any letter or digit, underscore, or hyphen. Complex names must be
surrounded by either a pair of quotes or double quotes. Some examples of
this are `$['name']` or `$["long-name"]`. This notation selects the
value of a JSON object by its name.

Index-style notation is a pair of `[]` with a series of digits in
between. Some examples of this are `$[0]` and `$[10]`. This notations
selects an element from a JSON array by its index.

In addition to selecting a single child, multiple children can be
selected using the bracket-style and index-style notations. This is
achieved by comma delimiting multiple selections. Some examples of this
are `$['name, alias']` or `[0, 1]`. This multiple selection behavior
returns an array with each of the values requested.