Add support for wildcards in children and indexes
Multiple children can now be selected using the wildcard character `*`.
The wildcard character can occur in the dot-style, bracket-style, and
index-style notations.

In the dot-style notation, the wildcard character *must* be the only
character following the dot and before the next selector. An example of
this is `$.store.*`. The wildcard will select every value for which
there is a key in a JSON object.

In the bracket-style notation, the wildcard character *must* be the only
character between either the pair of quotes or the pair of double
quotes. An example of this is `$.store['*']`. The wildcard will select
every value for which there is a key in a JSON object.

In the index-style notation, the wildcard character *must* be the only
character between the pair of brackets. An example of this is
`$.store[*]`. The wildcard will select every element in a JSON array.