Add support for deep wildcards
All children at all depths can now be selected using the deep wildcard
`..` syntax. The deep wildcard syntax is a special case of the dot-style
notation for children. The deep wildcard syntax is the logical
equivalent of a dot-style child with no name. The lack of the name
indicates that no children should be filtered out.

In the deep wildcard syntax, two dots are placed next to one another
with nothing in between. An example of this is `$.store..style`. This
syntax will select the value of the `style` key in any object below
`store` that has a `style` key. Another example is `$..*` which selects
absolutely every node in a JSON tree whether it is a JSON object, JSON
array, or value.