Direct Jackson support
Previously, the use of Jackson as the underlying JSON processing
implementation was hidden.  Since Jackson is clearly the most widely
used JSON implementation, it makes sense to expose its types through
the main JsonPath interface.  This allows the direct use of the
Jackson API both before and after JsonPath processing and eliminates
the need to convert back and forth between the right Jackson types and
basic strings.

This change is tracked in #3.