Travis CI improvements
After the first Travis CI run, various improvements needed to be made.
First, the name of the tests needed to be singularized so that they
would be picked up by Surefire by default.  Second the tests needed to
run against the Oracle JDK instead of Open JDK.  Third, the README.md
file was updated with the current status images.